requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
python-dotenv==0.14.0